#300: Change fixtures class name, package and contents.
